[React Weather] This creates components like Main,Nav, Weather(which is similar to Greeter/GreeterForm/GreeterMessage using the pass the function logic to pass the input from WeatherForm to Weather).
-> Use of open weather org api using axios library. axios.get(url)
is a promise based api which would return temperature on success case.

->The use of setState and getInitialState helps in state maintenance
and the render function of Weather can use this.state.<> to access
updated state variables.

-> cheap-eval-source-map-module helps enable source maps so that
bundle.js would be a one-one mapping of the source code instead of
the usual not so relevant bundle.js file. This introduces debugger;
keyword which would stop debugging similar to Ctrl+B .But it is
observed that only the first debugger instance works.

->isLoading is something that is introduced afresh.It is a boolean
variable which is set in getInitialState/setState and render function
renders according to the isLoading boolean variable.

->arrow functions is introduced in this batch. It is observed that
the difference between anonymous and an arrow function is the relevance
of the 'this' operator. anonymous function has its own 'this' operator
where as arrow function uses the 'this' operator of the parent.
As its already common knowledge, node arrowfunctions.jsx allows
us to execute js file directly in command line.

Signed-off-by: Srikanth Chevalam <srikanth.chevalam@gmail.com>